fix: format structured MCP tool results
- pass successful structuredContent to the model as pretty JSON
- keep tool error content unchanged when calls fail
- store structuredContent in details instead of the full MCP result
- fold MCP connection lifecycle tests into the main Pi spec
