+++ packages/pi/tests/main.spec.ts
@@ -1,4 +1,5 @@
 import { afterEach, describe, expect, mock, test } from 'bun:test'
+import { createMcpHandler, fromJsonSchema, McpServer } from '@modelcontextprotocol/server'
 
 type RegisteredTool = {
   name: string
@@ -47,6 +48,86 @@ const callBeforeAgentStart = async (
 
   if (!result) throw new Error('handler returned nothing')
   return result
+}
+
+/** Real in-process MCP server: the extension's transport fetch is wired straight to handler.fetch. */
+const createTestServer = () => {
+  let initializeCount = 0
+  let failNextToolCall = false
+
+  const handler = createMcpHandler(() => {
+    const server = new McpServer({ name: 'fixture', version: '1.0.0' })
+    server.registerTool(
+      'echo',
+      {
+        description: 'Echo the query back',
+        inputSchema: fromJsonSchema<{ query: string }>({
+          type: 'object',
+          properties: { query: { type: 'string' } },
+          required: ['query'],
+        }),
+      },
+      async ({ query }) => ({ content: [{ type: 'text' as const, text: `echo:${query}` }] }),
+    )
+    server.registerTool(
+      'structured-echo',
+      {
+        description: 'Echo the query back as structured content',
+        inputSchema: fromJsonSchema<{ query: string }>({
+          type: 'object',
+          properties: { query: { type: 'string' } },
+          required: ['query'],
+        }),
+        outputSchema: fromJsonSchema<{ echo: string }>({
+          type: 'object',
+          properties: { echo: { type: 'string' } },
+          required: ['echo'],
+        }),
+      },
+      async ({ query }) => {
+        const output = { echo: query }
+        return {
+          content: [{ type: 'text' as const, text: JSON.stringify(output) }],
+          structuredContent: output,
+        }
+      },
+    )
+    return server
+  })
+
+  const fetchHandler = async (url: string | URL, init?: RequestInit) => {
+    const request = new Request(String(url), init)
+    if (request.method === 'POST') {
+      const body = (await request.clone().json()) as { method?: string }
+      if (body.method === 'initialize') initializeCount += 1
+      if (body.method === 'tools/call' && failNextToolCall) {
+        failNextToolCall = false
+        throw new TypeError('fetch failed')
+      }
+    }
+    return handler.fetch(request)
+  }
+
+  return {
+    close: () => handler.close(),
+    failNextToolCall: () => {
+      failNextToolCall = true
+    },
+    initializeCount: () => initializeCount,
+    serverConfig: {
+      url: 'http://fixture.local/mcp',
+      authenticated: false,
+      fetch: fetchHandler,
+      promptGuidelines: ['fixture server'],
+    },
+  }
+}
+
+const findTool = (tools: RegisteredTool[], name: string) => {
+  const tool = tools.find((registeredTool) => registeredTool.name === name)
+  expect(tool).toBeDefined()
+  if (!tool) throw new Error(`${name} tool was not registered`)
+  return tool
 }
 
 const YDC_API_KEY = process.env.YDC_API_KEY ?? ''
@@ -100,7 +181,7 @@ describe('Pi extension', () => {
       expect(duplicates).toEqual([])
     })
 
-    test('passes MCP content text blocks to the model without JSON-wrapping the full result', async () => {
+    test('passes successful structured content to the model without JSON-wrapping the full result', async () => {
       const extension = await loadExtension()
       const { pi, tools } = createPiMock()
       process.env.YDC_API_KEY = YDC_API_KEY
@@ -112,10 +193,10 @@ describe('Pi extension', () => {
 
       const result = (await tool.execute('call-1', { query: 'OpenAI' })) as {
         content: Array<{ type: string; text: string }>
-        details: { structuredContent?: unknown }
+        details: unknown
       }
 
-      // Model-facing content is raw text blocks from the MCP server, not JSON.stringify(result)
+      // Model-facing content is structuredContent, not JSON.stringify(result)
       expect(result.content.length).toBeGreaterThan(0)
       expect(result.content.every((block) => block.type === 'text')).toBe(true)
       const firstBlock = result.content[0]
@@ -124,8 +205,7 @@ describe('Pi extension', () => {
       // The text must not be a JSON wrapper of the entire MCP response (which would include structuredContent)
       expect(firstBlock.text).not.toContain('structuredContent')
       expect(firstBlock.text).not.toMatch(/^\{"content":/)
-      // Full raw result (including structuredContent) is preserved in details for UI/logs
-      expect(result.details.structuredContent).toBeDefined()
+      expect(result.details).toBeDefined()
     })
 
     test('rejects invalid tool input before crossing the MCP boundary', async () => {
@@ -161,6 +241,122 @@ describe('Pi extension', () => {
       expect(result.systemPrompt).toContain('https://api.you.com/mcp?tools=you-finance')
       expect(result.systemPrompt).toContain('https://api.you.com/mcp')
       expect(result.systemPrompt).toContain('https://you.com/docs/_mcp/server')
+    })
+  })
+
+  describe('MCP connection lifecycle', () => {
+    test('pretty-prints structured content for successful tool calls', async () => {
+      const server = createTestServer()
+      try {
+        const extension = await loadExtension()
+        const { pi, tools } = createPiMock()
+
+        await extension(pi, [server.serverConfig])
+
+        const tool = findTool(tools, 'structured-echo')
+        const result = (await tool.execute('call-1', { query: 'one' })) as {
+          content: Array<{ text: string }>
+          details: unknown
+        }
+
+        expect(result.content[0]?.text).toBe('{\n  "echo": "one"\n}')
+        expect(result.details).toEqual({ echo: 'one' })
+      } finally {
+        await server.close()
+      }
+    })
+
+    test('preserves tool error text instead of formatting missing structured content', async () => {
+      const server = createTestServer()
+      try {
+        const extension = await loadExtension()
+        const { pi, tools } = createPiMock()
+
+        await extension(pi, [server.serverConfig])
+
+        const tool = findTool(tools, 'structured-echo')
+        const result = (await tool.execute('call-1', { query: 1 })) as {
+          content: Array<{ text: string }>
+          details: unknown
+        }
+
+        expect(result.content[0]?.text).toContain('Input validation error')
+        expect(result.details).toBeUndefined()
+      } finally {
+        await server.close()
+      }
+    })
+
+    test('reuses one MCP connection across tool executions', async () => {
+      const server = createTestServer()
+      try {
+        const extension = await loadExtension()
+        const { pi, tools } = createPiMock()
+
+        await extension(pi, [server.serverConfig])
+        // Tool discovery pays its own one-shot handshake; start counting from here.
+        const afterDiscovery = server.initializeCount()
+
+        const tool = findTool(tools, 'echo')
+        const first = (await tool.execute('call-1', { query: 'one' })) as { content: Array<{ text: string }> }
+        const second = (await tool.execute('call-2', { query: 'two' })) as { content: Array<{ text: string }> }
+
+        expect(first.content[0]?.text).toBe('echo:one')
+        expect(second.content[0]?.text).toBe('echo:two')
+        expect(server.initializeCount()).toBe(afterDiscovery + 1)
+      } finally {
+        await server.close()
+      }
+    })
+
+    test('reconnects and retries once when the connection drops', async () => {
+      const server = createTestServer()
+      try {
+        const extension = await loadExtension()
+        const { pi, tools } = createPiMock()
+
+        await extension(pi, [server.serverConfig])
+        const afterDiscovery = server.initializeCount()
+
+        const tool = findTool(tools, 'echo')
+        await tool.execute('call-1', { query: 'one' })
+        expect(server.initializeCount()).toBe(afterDiscovery + 1)
+
+        server.failNextToolCall()
+        const result = (await tool.execute('call-2', { query: 'two' })) as { content: Array<{ text: string }> }
+
+        expect(result.content[0]?.text).toBe('echo:two')
+        // The dropped pooled client was replaced: one new handshake, then success.
+        expect(server.initializeCount()).toBe(afterDiscovery + 2)
+      } finally {
+        await server.close()
+      }
+    })
+
+    test('closes pooled clients on session_shutdown', async () => {
+      const server = createTestServer()
+      try {
+        const extension = await loadExtension()
+        const { events, pi, tools } = createPiMock()
+
+        await extension(pi, [server.serverConfig])
+        const afterDiscovery = server.initializeCount()
+
+        const tool = findTool(tools, 'echo')
+        await tool.execute('call-1', { query: 'one' })
+        expect(server.initializeCount()).toBe(afterDiscovery + 1)
+
+        const shutdown = events.find((event) => event.eventName === 'session_shutdown')
+        expect(shutdown).toBeDefined()
+        if (!shutdown) throw new Error('session_shutdown handler was not registered')
+        await shutdown.handler()
+
+        await tool.execute('call-2', { query: 'two' })
+        // Shutdown dropped the pooled client: the next execution reconnects.
+        expect(server.initializeCount()).toBe(afterDiscovery + 2)
+      } finally {
+        await server.close()
+      }
     })
   })
 })
